Move uefi search to library.uefi folder
