+++ chipsec/modules/tools/uefi/scan_blocked.py
@@ -57,7 +57,7 @@ from chipsec.library.returncode import ModuleResult
 from chipsec.hal.common.spi_uefi import search_efi_tree, build_efi_model, EFIModuleType
 from chipsec.hal.common.uefi import UEFI
 from chipsec.hal.intel.spi import SPI, BIOS
-from chipsec.hal.common.uefi_search import check_match_criteria
+from chipsec.library.uefi.uefi_search import check_match_criteria
 from chipsec.library.file import read_file, get_main_dir
 
 TAGS = [MTAG_BIOS]
